requirements.txt
certifi==2020.6.20
cycler==0.10.0
future==0.18.2
kiwisolver==1.2.0
matplotlib==3.3.2
numpy==1.19.2
opt-einsum==3.3.0
Pillow==7.2.0
pyparsing==2.4.7
pyro-api==0.1.2
pyro-ppl==1.4.0
python-dateutil==2.8.1
six==1.15.0
torch==1.6.0
torchvision==0.7.0
tqdm==4.50.0
